Add file based locking to avoid race conditions in key generation

+++ src/rasenmaeher_api/jwtinit.py
@@ -4,7 +4,9 @@ import logging
 import os
 import asyncio
 from pathlib import Path
+import random
 
+import filelock
 from multikeyjwt import Issuer
 from multikeyjwt.keygen import generate_keypair
 
@@ -69,10 +71,8 @@ def check_jwt_init() -> bool:
     return check_public_keys()
 
 
-async def jwt_init() -> None:
-    """If needed: Create keypair"""
-    if check_jwt_init():
-        return
+def _jwt_init_keypath() -> Path:
+    """resolve the path for the private key"""
     keypath: Union[Path, Optional[str]] = os.environ.get("JWT_PRIVKEY_PATH")
     if keypath:
         keypath = Path(str(keypath))
@@ -84,18 +84,11 @@ async def jwt_init() -> None:
         keypath = DEFAULT_KEY_PATH
     if not keypath.parent.exists():
         keypath.parent.mkdir(parents=True, mode=0o760)
-    if keypass := os.environ.get("JWT_PRIVKEY_PASS"):
-        LOGGER.info("Private key password defined in ENV, going to use it")
+    return keypath
 
-    LOGGER.info("Running keygen in executor")
-    genpubpath: Optional[Path] = None
-    genprivpath: Optional[Path] = None
-    genprivpath, genpubpath = await asyncio.get_running_loop().run_in_executor(None, generate_keypair, keypath, keypass)
-    if not genprivpath or not genprivpath.exists():
-        raise RuntimeError("Returned private key does not exist!")
-    if not genpubpath or not genpubpath.exists():
-        raise RuntimeError("Returned private key does not exist!")
 
+def _jwt_init_pubkeypath(genpubpath: Path) -> Path:
+    """resolve the path for the public key"""
     pubkeypath: Union[Path, Optional[str]] = os.environ.get("JWT_PUBKEY_PATH")
     LOGGER.debug("initial pubkeypath={}".format(pubkeypath))
     if pubkeypath:
@@ -106,9 +99,47 @@ async def jwt_init() -> None:
     else:
         pubkeypath = DEFAULT_PUB_PATH
     LOGGER.debug("final pubkeypath={}".format(pubkeypath))
-
     if not pubkeypath.parent.exists():
         pubkeypath.parent.mkdir(parents=True)
+    return pubkeypath
+
+
+async def jwt_init() -> None:
+    """If needed: Create keypair"""
+    if check_jwt_init():
+        return
+    keypath = _jwt_init_keypath()
+
+    genpubpath: Optional[Path] = None
+    genprivpath: Optional[Path] = None
+    lockpath = keypath.with_suffix(".lock")
+    # Random sleep to avoid race conditions on these file accesses
+    await asyncio.sleep(random.random() * 3.0)  # nosec
+    lock = filelock.FileLock(lockpath)
+    try:
+        lock.acquire(timeout=0.0)
+        # Check the privkey again to avoid overwriting.
+        if keypath.exists():
+            return None
+        LOGGER.info("Running keygen in executor")
+        if keypass := os.environ.get("JWT_PRIVKEY_PASS"):
+            LOGGER.info("Private key password defined in ENV, going to use it")
+        genprivpath, genpubpath = await asyncio.get_running_loop().run_in_executor(
+            None, generate_keypair, keypath, keypass
+        )
+    except filelock.Timeout:
+        LOGGER.warning("Someone has already locked {}".format(lockpath))
+        LOGGER.debug("Sleeping for ~5s and then recursing")
+        await asyncio.sleep(5.0 + random.random())  # nosec
+        return await jwt_init()
+    finally:
+        lock.release()
+    if not genprivpath or not genprivpath.exists():
+        raise RuntimeError("Returned private key does not exist!")
+    if not genpubpath or not genpubpath.exists():
+        raise RuntimeError("Returned private key does not exist!")
+    pubkeypath = _jwt_init_pubkeypath(genpubpath)
+
     LOGGER.debug("Copy generated pubkey to {}".format(pubkeypath))
     pubkeypath.write_bytes(genpubpath.read_bytes())
     # Make sure KRAFTWERK public keys get copied
